refactor: extract message-menu-bar's logic into custom hooks
